Make the sphere hit function a common one and call from static sphere and moving sphere
