+++ raytracer/src/objects/sphere/static_sphere.rs
@@ -56,42 +56,7 @@ impl Hittable for Sphere {
     /// ![Ray-Sphere intersection results][raysphere]
     #[embed_doc_image("raysphere", "doc_images/ray_sphere_intersection_results.jpg")]
     fn hit(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
-        let oc = r.origin() - self.center;
-        let a = r.direction().length_squared();
-        let half_b = oc.dot(&r.direction());
-        let c = oc.length_squared() - self.radius * self.radius;
-
-        let discriminant = half_b * half_b - a * c;
-        if discriminant < 0.0 {
-            return None;
-        }
-
-        let sqrtd = f64::sqrt(discriminant);
-
-        // Find the nearest root that lies in the acceptable range
-        let mut root = (-half_b - sqrtd) / a;
-        if root < t_min || root > t_max {
-            root = (-half_b + sqrtd) / a;
-            if root < t_min || root > t_max {
-                return None;
-            }
-        }
-
-        let p = r.at(root);
-        let outward_normal = (p - self.center) / self.radius;
-        let (u, v) = common::get_sphere_uv(&outward_normal);
-        let mut hit_rec = HitRecord {
-            t: root,
-            p,
-            normal: Vec3::new(0.0, 0.0, 0.0),
-            front_face: false,
-            mat: self.material.clone(),
-            u,
-            v,
-        };
-        hit_rec.set_face_normal(r, &outward_normal);
-
-        Some(hit_rec)
+        common::hit(r, t_min, t_max, &|pt| pt - self.center, &|_|self.center, self.radius, self.material.clone())
     }
 
     fn bounding_box(&self, _time0: f64, _time1: f64) -> Option<AABB> {
